preparing api
