+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "start": "node index.js",
     "format": "prettier --write lib/**/*.js test/**/*.js *.js --single-quote --print-width 120",
-    "test": "mocha --timeout 12000"
+    "test": "mocha --timeout 12000 test/**/*.test.js"
   },
   "husky": {
     "hooks": {
@@ -39,8 +39,10 @@
   },
   "license": "MIT",
   "dependencies": {
+    "body-parser": "1.19.0",
     "lowdb": "1.0.0",
     "request-x-ray": "0.1.4",
+    "restana": "4.0.8",
     "slack": "11.0.2",
     "tg-yarl": "1.3.0",
     "x-ray": "2.3.4"
@@ -48,8 +50,8 @@
   "devDependencies": {
     "chai": "4.2.0",
     "eslint": "6.8.0",
-    "husky": "4.2.3",
     "eslint-config-prettier": "6.10.0",
+    "husky": "4.2.3",
     "lint-staged": "10.0.7",
     "mocha": "7.0.1",
     "prettier": "1.19.1",
